Your commit message doesn't match the code differences. Here's the correct commit message:
Refactor main.py and import scipy

Updated main.py to import the scipy library and encapsulated the usage part in the main function. This change not only assures the given code is being executed only when the script is running directly, but also enables the use of mathematical algorithms provided by scipy if necessary.